+++ config/jest/setupTests.js
@@ -1,0 +1,9 @@
+/* eslint-disable */
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+global.requestAnimationFrame = function(callback) {
+  setTimeout(callback, 0);
+};
+
+configure({ adapter: new Adapter() });