+++ src/actions.js
@@ -1,0 +1,6 @@
+import * as counterActions from './containers/Counter/actions';
+
+export default Object.assign(
+  {},
+  counterActions,
+);