+++ src/index.jsx
@@ -1,7 +1,32 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { ConnectedRouter } from 'react-router-redux';
+import createHistory from 'history/createBrowserHistory';
+import { AppContainer } from 'react-hot-loader';
 
-ReactDOM.render(
-  <h1>Hello World</h1>,
-  document.getElementById('root'),
-);
+import App from './App';
+import configureStore from './store';
+
+const history = createHistory();
+const store = configureStore({}, history);
+
+const render = (Component) => {
+  ReactDOM.render(
+    <Provider store={store}>
+      <ConnectedRouter history={history}>
+        <AppContainer>
+          <Component />
+        </AppContainer>
+      </ConnectedRouter>
+    </Provider>,
+    document.getElementById('root'),
+  );
+};
+render(App);
+
+if (module.hot) {
+  module.hot.accept('./App', () => {
+    render(App);
+  });
+}