+++ src/reducers.js
@@ -1,0 +1,5 @@
+import counterReducer from './containers/Counter/reducers';
+
+export default {
+  counter: counterReducer,
+};