+++ src/routes.js
@@ -1,0 +1,10 @@
+import Counter from './containers/Counter';
+
+const routes = [
+  {
+    path: '/',
+    component: Counter,
+  },
+];
+
+export default routes;