+++ CHANGELOG.md
@@ -7,3 +7,8 @@ All notable changes to this project will be documented in this file.
 ## [0.1.1] - 25 October 2017
 ### Fixed
 - Fix lint-staged doesn't match staged files
+
+## [0.2.0] - 18 November 2017
+### Added
+- Added package redux, react-hot-loader, router, jest, enzyme
+- Added Counter, routes as example